+++ server/scripts/test_connection.js
@@ -1,5 +1,5 @@
-let {User, initUser} = require('./models/users')
-let {getSequelize} = require('./connectionFactory')
+let {User, initUser} = require('../models/users')
+let {getSequelize} = require('../models/connectionFactory')
 
 async function setupSequelize(env) {
     const sequelize = getSequelize(env);
@@ -8,10 +8,9 @@ async function setupSequelize(env) {
     await User.sync()
 
 
-    const user = await User.create({id: 1, name: 'bob', email: 'bob@gmail.com', role: 'standard'});
+    const user = await User.create({id: 18898, name: 'bob', email: 'bob8@gmail.com', role: 'standard'});
     console.log(user)
     console.log(await User.count())
-
 }
 
 setupSequelize('development')
